Solving dependencies issues, geolocation is now working with Spark 4.0 and other upgraded libraries!!"

+++ scala/src/main/scala/geo/GeoElement.scala
@@ -17,7 +17,7 @@ case class GeoElement(
   , adm2_code:String
   , adm1_code:String
   , country_code:String
-  , population:Int
+  , population:Long
   , geo_type:String
   , time_zone:String
   , longitude:Double
